# examples/output-sqlite/openapi.yaml
...
openapi: 3.0.3
info:
  title: InventoryPlatform
  version: 1.0.0
  description: Generated by MarrowScript compiler
servers:
  - 
    url: "http://localhost:3000"
paths:
  /players:
    get:
      summary: List Player
      operationId: listPlayer
      tags:
        - Player
      parameters:
        - 
          name: page
          in: query
          schema:
            type: integer
            default: 1
        - 
          name: page_size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        200:
          description: List of Player
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Player"
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
        401:
          description: Unauthorized
    post:
      summary: Create Player
      operationId: createPlayer
      tags:
        - Player
      security:
        - 
          BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Player"
      responses:
        200:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        400:
          description: Bad request
        401:
          description: Unauthorized
        422:
          description: Precondition failed
  /players/{id}:
    get:
      summary: Get Player
      operationId: getPlayer
      tags:
        - Player
      parameters:
        - 
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - 
          BearerAuth: []
      responses:
        200:
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        400:
          description: Not found
        401:
          description: Unauthorized
    put:
      summary: Update Player
      operationId: updatePlayer
      tags:
        - Player
      parameters:
        - 
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - 
          BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Player"
      responses:
        200:
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        400:
          description: Bad request
        401:
          description: Unauthorized
        422:
          description: Precondition failed
    delete:
      summary: Delete Player
      operationId: deletePlayer
      tags:
        - Player
      parameters:
        - 
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - 
          BearerAuth: []
      responses:
        200:
          description: Deleted
        400:
          description: Not found
        401:
          description: Unauthorized
  /players/create-trade:
    post:
      summary: create_trade on Player
      operationId: create_tradePlayer
      tags:
        - Player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                from_id:
                  type: string
                  format: uuid
                  description: UUID of an existing Player row (alternative to inline body)
                id:
                  type: string
                  format: uuid
                created_at:
                  type: string
                  format: date-time
                updated_at:
                  type: string
                  format: date-time
                username:
                  type: string
                email:
                  type: string
                score:
                  type: integer
                level:
                  type: integer
                inventory:
                  type: array
                  items:
                    type: string
                to_id:
                  type: string
                  format: uuid
                  description: UUID of an existing Player row (alternative to inline body)
                items:
                  type: array
                  items: {}
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  action:
                    type: string
        400:
          description: Bad request
        401:
          description: Unauthorized
        404:
          description: Not found
        422:
          description: Precondition failed
      security:
        - 
          BearerAuth: []
  /players/accept-trade:
    post:
      summary: accept_trade on Player
      operationId: accept_tradePlayer
      tags:
        - Player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                trade_id:
                  type: string
                  format: uuid
                  description: UUID of an existing TradeOffer row (alternative to inline body)
                acceptor_id:
                  type: string
                  format: uuid
                  description: UUID of an existing Player row (alternative to inline body)
                id:
                  type: string
                  format: uuid
                created_at:
                  type: string
                  format: date-time
                updated_at:
                  type: string
                  format: date-time
                username:
                  type: string
                email:
                  type: string
                score:
                  type: integer
                level:
                  type: integer
                inventory:
                  type: array
                  items:
                    type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  action:
                    type: string
        400:
          description: Bad request
        401:
          description: Unauthorized
        404:
          description: Not found
        422:
          description: Precondition failed
      security:
        - 
          BearerAuth: []
  /players/award-xp:
    post:
      summary: award_xp on Player
      operationId: award_xpPlayer
      tags:
        - Player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                player_id:
                  type: string
                  format: uuid
                  description: UUID of an existing Player row (alternative to inline body)
                id:
                  type: string
                  format: uuid
                created_at:
                  type: string
                  format: date-time
                updated_at:
                  type: string
                  format: date-time
                username:
                  type: string
                email:
                  type: string
                score:
                  type: integer
                level:
                  type: integer
                inventory:
                  type: array
                  items:
                    type: string
                amount:
                  type: integer
                  minimum: 0
              required:
                - amount
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  action:
                    type: string
        400:
          description: Bad request
        401:
          description: Unauthorized
        404:
          description: Not found
        422:
          description: Precondition failed
      security:
        - 
          BearerAuth: []
  /players/consume-item:
    post:
      summary: consume_item on Player
      operationId: consume_itemPlayer
      tags:
        - Player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                player_id:
                  type: string
                  format: uuid
                  description: UUID of an existing Player row (alternative to inline body)
                id:
                  type: string
                  format: uuid
                created_at:
                  type: string
                  format: date-time
                updated_at:
                  type: string
                  format: date-time
                username:
                  type: string
                email:
                  type: string
                score:
                  type: integer
                level:
                  type: integer
                inventory:
                  type: array
                  items:
                    type: string
                item_id:
                  type: string
                  format: uuid
                  description: UUID of an existing Item row (alternative to inline body)
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  action:
                    type: string
        400:
          description: Bad request
        401:
          description: Unauthorized
        404:
          description: Not found
        422:
          description: Precondition failed
      security:
        - 
          BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Player:
      type: object
      properties:
        id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        username:
          type: string
        email:
          type: string
        score:
          type: integer
        level:
          type: integer
        inventory:
          type: array
          items:
            type: string
...