# examples/output-sqlite/src/routes/player.ts
// Generated by MarrowScript compiler. DO NOT EDIT.
// Service: PlayerService
import { Router, Request, Response } from "express";
import { v4 as uuid } from "uuid";
import { query, queryOne, execute, pool } from "../db";
import { eventBus } from "../events";
import { requireAuth, AuthContext } from "../auth";
import rateLimit from "express-rate-limit";
import { auditLog } from "../audit";
import { logger } from "../logger";
import { counter } from "../metrics";
import { PlayerCreateSchema, PlayerUpdateSchema } from "../schemas";
import * as __algorithms from "../algorithms";
const { shortestPath, topologicalSort, binarySearch, bipartiteMatching, roundRobin, weightedAverage, percentile, rankBy, consistentHash } = __algorithms as any;

// User-defined functions referenced in effects — implement these or use extension_point
declare function calculate_level(...args: any[]): any;

import { transitionPlayer, PLAYER_INITIAL } from "../state_machines/player";

export const playersRouter = Router();

// Rate limiter from policy declaration
const __routeRateLimit = rateLimit({ windowMs: 60000, max: 100, standardHeaders: true, legacyHeaders: false });

// CREATE
playersRouter.post("/", __routeRateLimit, requireAuth, async (req: Request, res: Response) => {
  // Validate the request body against the generated Zod schema.
  const __parsed = PlayerCreateSchema.safeParse(req.body);
  if (!__parsed.success) {
    return res.status(400).json({ error: { code: "VALIDATION_FAILED", message: "Invalid request body", issues: __parsed.error.flatten() } });
  }
  const __body = __parsed.data;
  try {
    const id = uuid();
    const { username, email, score, level, inventory } = __body as any;
    const state = PLAYER_INITIAL;
    const sql = `INSERT INTO players (id, username, email, score, level, inventory, state) VALUES ($1, $2, $3, $4, $5, $6, $7) RETURNING *`;
    const rows = await query(sql, [id, username, email, score, level, inventory, state]);
    counter("entity.created", { entity: "Player" });
    res.status(201).json(rows[0]);
  } catch (e: any) {
    res.status(400).json({ error: { code: "CREATE_FAILED", message: e.message } });
  }
});

// READ
playersRouter.get("/:id", __routeRateLimit, requireAuth, async (req: Request, res: Response) => {
  try {
    const row = await queryOne(`SELECT * FROM players WHERE id = $1`, [req.params.id]);
    if (!row) return res.status(404).json({ error: { code: "NOT_FOUND", message: "Not found" } });
    res.json(row);
  } catch (e: any) {
    res.status(500).json({ error: { code: "DB_ERROR", message: e.message } });
  }
});

// LIST
playersRouter.get("/", __routeRateLimit, requireAuth, async (req: Request, res: Response) => {
  try {
    const page = parseInt(req.query.page as string) || 1;
    const pageSize = Math.min(parseInt(req.query.page_size as string) || 50, 100);
    const offset = (page - 1) * pageSize;
    const rows = await query(`SELECT * FROM players ORDER BY created_at DESC LIMIT $1 OFFSET $2`, [pageSize, offset]);
    const [{ count }] = await query(`SELECT COUNT(*) as count FROM players`);
    res.json({ items: rows, total: parseInt(count), page, page_size: pageSize });
  } catch (e: any) {
    res.status(500).json({ error: { code: "DB_ERROR", message: e.message } });
  }
});

// UPDATE
const __playersUpdatable = new Set<string>(["username","email","score","level","inventory","state"]);
playersRouter.put("/:id", __routeRateLimit, requireAuth, async (req: Request, res: Response) => {
  // 1. Validate body shape and types via Zod (UpdateSchema is partial).
  const __parsed = PlayerUpdateSchema.safeParse(req.body);
  if (!__parsed.success) {
    return res.status(400).json({ error: { code: "VALIDATION_FAILED", message: "Invalid request body", issues: __parsed.error.flatten() } });
  }
  // 2. Reject unknown columns (defense in depth — Zod would already strip them).
  const __unknown = Object.keys(req.body || {}).filter(k => !__playersUpdatable.has(k));
  if (__unknown.length > 0) {
    return res.status(400).json({ error: { code: "UNKNOWN_FIELDS", message: `Unknown fields: ${__unknown.join(", ")}`, fields: __unknown } });
  }
  // 3. Use the Zod-parsed object as the update set.
  const fields: Record<string, unknown> = __parsed.data as Record<string, unknown>;
  if (Object.keys(fields).length === 0) {
    return res.status(400).json({ error: { code: "NO_FIELDS", message: "No updatable fields supplied" } });
  }
  // State machine enforcement
  if (fields.state !== undefined) {
    const current = await queryOne<{ state: string }>(`SELECT state FROM players WHERE id = $1`, [req.params.id]);
    if (!current) return res.status(404).json({ error: { code: "NOT_FOUND", message: "Not found" } });
    // Find the trigger for this state transition
    const trigger = `${current.state}_to_${fields.state}`;
    const tr = transitionPlayer(current.state as any, trigger);
    if (!tr.ok) return res.status(422).json({ error: { code: "INVALID_TRANSITION", message: `Cannot transition from ${current.state} to ${fields.state}` } });
  }
  const keys = Object.keys(fields);
  const sets = keys.map((k, i) => `${k} = $${i + 2}`).join(", ");
  const values = keys.map(k => fields[k]);
  const sql = `UPDATE players SET ${sets}, updated_at = NOW() WHERE id = $1 RETURNING *`;
  const rows = await query(sql, [req.params.id, ...values]);
  if (rows.length === 0) return res.status(404).json({ error: { code: "NOT_FOUND", message: "Not found" } });
  res.json(rows[0]);
});

// DELETE
playersRouter.delete("/:id", __routeRateLimit, requireAuth, async (req: Request, res: Response) => {
  try {
    const count = await execute(`DELETE FROM players WHERE id = $1`, [req.params.id]);
    if (count === 0) return res.status(404).json({ error: { code: "NOT_FOUND", message: "Not found" } });
    res.status(204).send();
  } catch (e: any) {
    res.status(500).json({ error: { code: "DB_ERROR", message: e.message } });
  }
});

// CAPABILITY: create_trade [transactional]
playersRouter.post("/create-trade", __routeRateLimit, requireAuth, auditLog("create_trade", "Player"), async (req: Request, res: Response) => {
  const auth: AuthContext = (req as any).auth;
  const __client = await pool.connect();
  try {
    await __client.query("BEGIN");
    const { items } = req.body;

    // Fetch entities
    const from = await queryOne(`SELECT * FROM players WHERE id = $1`, [req.body.from_id || req.params.id]);
    if (!from) {
      return res.status(404).json({ error: { code: "NOT_FOUND", message: "from not found" } });
    }
    const to = await queryOne(`SELECT * FROM players WHERE id = $1`, [req.body.to_id || req.params.id]);
    if (!to) {
      return res.status(404).json({ error: { code: "NOT_FOUND", message: "to not found" } });
    }

    // Preconditions
    if (!(from?.state === "active")) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "from.state == \\\"active\\\"" } });
    }
    if (!(to?.state === "active")) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "to.state == \\\"active\\\"" } });
    }
    if (!(from !== to)) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "from != to" } });
    }
    if (!(items?.size >= 1)) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "items.size >= 1" } });
    }
    if (!(items?.size <= 10)) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "items.size <= 10" } });
    }

    // Effects (applied in declaration order)
    // EFFECT: items.state = "reserved"
    // TODO: Implement this effect manually

    // Emit events
    await eventBus.publish("TradeOfferCreated", { from_id: from?.id, to_id: to?.id, timestamp: new Date().toISOString(), actor_id: auth.actor_id }, "PlayerService", auth.trace_id, __client);

    res.json({ ok: true, action: "create_trade", entity: from });
    await __client.query("COMMIT");
  } catch (e: any) {
    await __client.query("ROLLBACK");
    if (!res.headersSent) {
      res.status(400).json({ error: { code: "CAPABILITY_FAILED", message: e.message } });
    } else {
      logger.error("late_error_after_response", { event: "create_trade", status: "rejected", metadata: { error: String(e?.message || e).slice(0, 256) } });
    }
  } finally {
    __client.release();
  }
});

// CAPABILITY: accept_trade [transactional]
playersRouter.post("/accept-trade", __routeRateLimit, requireAuth, auditLog("accept_trade", "Player"), async (req: Request, res: Response) => {
  const auth: AuthContext = (req as any).auth;
  const __client = await pool.connect();
  try {
    await __client.query("BEGIN");
    // Fetch entities
    const trade = await queryOne(`SELECT * FROM trade_offers WHERE id = $1`, [req.body.trade_id || req.params.id]);
    if (!trade) {
      return res.status(404).json({ error: { code: "NOT_FOUND", message: "trade not found" } });
    }
    const acceptor = await queryOne(`SELECT * FROM players WHERE id = $1`, [req.body.acceptor_id || req.params.id]);
    if (!acceptor) {
      return res.status(404).json({ error: { code: "NOT_FOUND", message: "acceptor not found" } });
    }

    // Preconditions
    if (!(trade?.status === "pending")) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "trade.status == \\\"pending\\\"" } });
    }
    if (!(trade?.to_player === acceptor?.id)) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "trade.to_player == acceptor.id" } });
    }
    if (!(acceptor?.state === "active")) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "acceptor.state == \\\"active\\\"" } });
    }
    if (!(trade?.expires_at > new Date())) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "trade.expires_at > now()" } });
    }

    // Effects (applied in declaration order)
    // EFFECT: trade.status = "accepted"
    // TODO: Implement this effect manually
    // EFFECT: trade.offered_items.owner_id = trade.to_player
    // TODO: Implement this effect manually
    // EFFECT: trade.requested_items.owner_id = trade.from_player
    // TODO: Implement this effect manually

    // Emit events
    await eventBus.publish("ItemTraded", { trade_id: trade?.id, acceptor_id: acceptor?.id, timestamp: new Date().toISOString(), actor_id: auth.actor_id }, "PlayerService", auth.trace_id, __client);

    res.json({ ok: true, action: "accept_trade", entity: trade });
    await __client.query("COMMIT");
  } catch (e: any) {
    await __client.query("ROLLBACK");
    if (!res.headersSent) {
      res.status(400).json({ error: { code: "CAPABILITY_FAILED", message: e.message } });
    } else {
      logger.error("late_error_after_response", { event: "accept_trade", status: "rejected", metadata: { error: String(e?.message || e).slice(0, 256) } });
    }
  } finally {
    __client.release();
  }
});

// CAPABILITY: award_xp
playersRouter.post("/award-xp", __routeRateLimit, requireAuth, auditLog("award_xp", "Player"), async (req: Request, res: Response) => {
  const auth: AuthContext = (req as any).auth;
  try {
    // sync: eventual — effects applied, events queued via outbox
    const { amount } = req.body;

    // Fetch entities
    const player = await queryOne(`SELECT * FROM players WHERE id = $1`, [req.body.player_id || req.params.id]);
    if (!player) {
      return res.status(404).json({ error: { code: "NOT_FOUND", message: "player not found" } });
    }

    // Preconditions
    if (!(player?.state === "active")) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "player.state == \\\"active\\\"" } });
    }
    if (!(amount > 0)) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "amount > 0" } });
    }
    if (!(amount <= 10000)) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "amount <= 10000" } });
    }

    // Effects (applied in declaration order)
    const __effect_0 = await query(`UPDATE players SET score = score + $1, updated_at = NOW() WHERE id = $2 RETURNING *`, [amount, req.body.player_id || req.params.id]);
    if (!__effect_0 || __effect_0.length === 0) {
      throw new Error("Effect failed: player.score += amount");
    }
    const __effect_1 = await query(`UPDATE players SET level = $1, updated_at = NOW() WHERE id = $2 RETURNING *`, [calculate_level(player?.score + amount), req.body.player_id || req.params.id]);
    if (!__effect_1 || __effect_1.length === 0) {
      throw new Error("Effect failed: player.level = calculate_level((player.score + amount))");
    }

    // Emit events
    await eventBus.publish("PlayerLeveledUp", { player_id: player?.id, timestamp: new Date().toISOString(), actor_id: auth.actor_id }, "PlayerService", auth.trace_id);

    res.json({ ok: true, action: "award_xp", entity: player });
  } catch (e: any) {
    if (!res.headersSent) {
      res.status(400).json({ error: { code: "CAPABILITY_FAILED", message: e.message } });
    } else {
      logger.error("late_error_after_response", { event: "award_xp", status: "rejected", metadata: { error: String(e?.message || e).slice(0, 256) } });
    }
  }
});

// CAPABILITY: consume_item [transactional]
playersRouter.post("/consume-item", __routeRateLimit, requireAuth, auditLog("consume_item", "Player"), async (req: Request, res: Response) => {
  const auth: AuthContext = (req as any).auth;
  const __client = await pool.connect();
  try {
    await __client.query("BEGIN");
    // Fetch entities
    const player = await queryOne(`SELECT * FROM players WHERE id = $1`, [req.body.player_id || req.params.id]);
    if (!player) {
      return res.status(404).json({ error: { code: "NOT_FOUND", message: "player not found" } });
    }
    const item = await queryOne(`SELECT * FROM items WHERE id = $1`, [req.body.item_id || req.params.id]);
    if (!item) {
      return res.status(404).json({ error: { code: "NOT_FOUND", message: "item not found" } });
    }

    // Preconditions
    if (!(player?.state === "active")) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "player.state == \\\"active\\\"" } });
    }
    if (!(item?.owner_id === player?.id)) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "item.owner_id == player.id" } });
    }
    if (!(item?.state === "available")) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "item.state == \\\"available\\\"" } });
    }
    if (!(item?.quantity > 0)) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "item.quantity > 0" } });
    }

    // Effects (applied in declaration order)
    const __effect_0 = await query(`UPDATE items SET quantity = quantity - $1, updated_at = NOW() WHERE id = $2 RETURNING *`, [1, req.body.item_id || req.params.id]);
    if (!__effect_0 || __effect_0.length === 0) {
      throw new Error("Effect failed: item.quantity -= 1");
    }

    res.json({ ok: true, action: "consume_item", entity: player });
    await __client.query("COMMIT");
  } catch (e: any) {
    await __client.query("ROLLBACK");
    if (!res.headersSent) {
      res.status(400).json({ error: { code: "CAPABILITY_FAILED", message: e.message } });
    } else {
      logger.error("late_error_after_response", { event: "consume_item", status: "rejected", metadata: { error: String(e?.message || e).slice(0, 256) } });
    }
  } finally {
    __client.release();
  }
});